+++ src/main/java/ma/vi/esql/lookup/LookupExtension.java
@@ -25,23 +25,15 @@ public class LookupExtension implements Extension {
   @Override
   public void init(Database db, Configuration config) {
     this.db = db;
-    this.schema = config.get("schema", "_lookup");
-    log.log(INFO, "Creating lookup tables in " + db + " in schema " + schema);
+    log.log(INFO, "Creating lookup tables in " + db + " in schema _lookup");
     try (EsqlConnection c = db.esql()) {
       ///////////////////////////////////////////////////////////////////////////
       // Create lookup tables
       ///////////////////////////////////////////////////////////////////////////
       c.exec("""
-             create table %1$s.Lookup drop undefined({
+             create table _lookup.Lookup drop undefined({
                name: 'Lookup',
-               description: 'A named table of values',
-               dependents: {
-                 links: {
-                   type:        '%1$s.LookupLink',
-                   referred_by: 'source_lookup_id',
-                   label:       'Lookup Links'
-                 }
-               }
+               description: 'A named table of values'
              }
 
              _id         uuid not null,
@@ -50,7 +42,6 @@ public class LookupExtension implements Extension {
              _can_edit   bool not null default true,
 
              name string not null {
-               validate_unique: true,
                mask: 'Iiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiii'
              },
              display_name string not null,
@@ -58,54 +49,35 @@ public class LookupExtension implements Extension {
              "group"      string,
 
              primary key(_id),
-             unique(name))""".formatted(schema));
+             unique(name))""");
 
       c.exec("""
-          create table %1$s.LookupLink drop undefined({
-            name: 'Lookup Link',
-            description: 'The definition of links between values of lookup tables which are used for searching data by associations and for aggregating data in reports'
-          }
-
-          _id         uuid not null,
-          _version    long not null default 0,
-          _can_delete bool not null default true,
-          _can_edit   bool not null default true,
-
-          name string not null {
-            validate_unique: true,
-            mask: 'Iiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiii',
-            description: 'Start with a letter, follow by letters or digits'
-          },
-
-          display_name string not null,
-
-          source_lookup_id uuid not null { show: false },
-
-          target_lookup_id uuid not null {
-            label:         'Target',
-            value_table:   '%1$s.Lookup',
-            value_id:      '_id',
-            value_label:   'name',
-            show_value_as: joinlabel(target_lookup_id, '_id', 'name', '%1$s.Lookup')
-          },
-
-          primary key(_id),
-          unique(name),
-          foreign key(source_lookup_id) references %1$s.Lookup(_id),
-          foreign key(target_lookup_id) references %1$s.Lookup(_id))""".formatted(schema));
+             create table _lookup.LookupLink drop undefined({
+               name: 'Lookup Link',
+               description: 'The definition of links between values of lookup tables which are used for searching data by associations and for aggregating data in reports'
+             }
+    
+             _id         uuid not null,
+             _version    long not null default 0,
+             _can_delete bool not null default true,
+             _can_edit   bool not null default true,
+    
+             source_lookup_id uuid not null { show: false },
+    
+             target_lookup_id uuid not null {
+               link_table:   '_lookup.Lookup',
+               link_code:    '_id',
+               link_label:   'name'
+             },
+    
+             primary key(_id),
+             foreign key(source_lookup_id) references _lookup.Lookup(_id),
+             foreign key(target_lookup_id) references _lookup.Lookup(_id))""");
 
       c.exec("""
-             create table %1$s.LookupValue drop undefined({
+             create table _lookup.LookupValue drop undefined({
                name: 'Lookup Value',
-               description: 'The values in a lookup table',
-               validate_unique: [['lookup_id', 'code', 'lang']],
-               dependents: {
-                 links: {
-                   type: '%1$s.LookupValueLink',
-                   referred_by: 'source_value_id',
-                   label: 'Lookup Value Links'
-                 }
-               }
+               description: 'The values in a lookup table'
              }
 
              _id         uuid not null,
@@ -113,7 +85,7 @@ public class LookupExtension implements Extension {
              _can_delete bool not null default true,
              _can_edit   bool not null default true,
 
-             lookup_id uuid not null { show: false },
+             lookup_id   uuid not null { show: false },
              
              code        string not null,
              alt_code1   string,
@@ -128,10 +100,10 @@ public class LookupExtension implements Extension {
 
              primary key(_id),
              unique(lookup_id, code, lang),
-             foreign key(lookup_id) references %1$s.Lookup(_id) on delete cascade on update cascade)""".formatted(schema));
+             foreign key(lookup_id) references _lookup.Lookup(_id) on delete cascade on update cascade)""");
 
       c.exec("""
-             create table %1$s.LookupValueLink drop undefined({
+             create table _lookup.LookupValueLink drop undefined({
                name: 'Lookup Value Link',
                description: 'Links between values of lookup tables, used primarily for searching data by associations and for aggregating data in reports'
              }
@@ -146,8 +118,8 @@ public class LookupExtension implements Extension {
              target_value_id uuid   not null,
 
              primary key(_id),
-             foreign key(source_value_id) references %1$s.LookupValue(_id),
-             foreign key(target_value_id) references %1$s.LookupValue(_id))""".formatted(schema));
+             foreign key(source_value_id) references _lookup.LookupValue(_id),
+             foreign key(target_value_id) references _lookup.LookupValue(_id))""");
     }
 
     /*
@@ -157,8 +129,8 @@ public class LookupExtension implements Extension {
     // labels functions and macros
     ////////////////////////////////
     Structure structure = db.structure();
-    structure.function(new LookupLabelFunction(schema));
-    structure.function(new LookupLabel(schema));
+    structure.function(new LookupLabelFunction());
+    structure.function(new LookupLabel());
     structure.function(new JoinLabel());
 
     /*
@@ -169,10 +141,10 @@ public class LookupExtension implements Extension {
       try (Connection c = db.pooledConnection()) {
         // lookup label with no links
         c.createStatement().executeUpdate("""
-            create or replace function "%1$s".lookup_label(code text,
-                                                          lookup text,
-                                                          show_code boolean,
-                                                          show_label boolean) returns text as $$
+            create or replace function _lookup.lookup_label(code text,
+                                                              lookup text,
+                                                              show_code boolean,
+                                                              show_label boolean) returns text as $$
                 select case when coalesce(show_code, false)=coalesce(show_label, false)
                             then v.code || ' - ' || v.label
             
@@ -181,19 +153,18 @@ public class LookupExtension implements Extension {
             
                             else v.label
                        end
-                  from "%1$s"."LookupValue" v
-                  join "%1$s"."Lookup"      l on v.lookup_id=l._id
+                  from _lookup."LookupValue" v
+                  join _lookup."Lookup"      l on v.lookup_id=l._id
                  where l.name=$2 and v.code=$1;
-            $$ language sql immutable;
-            """.formatted(schema));
+            $$ language sql immutable;""");
 
         // lookup label with variable number of links
         c.createStatement().executeUpdate("""
-            create or replace function "%1$s".lookup_label(code text,
-                                                          lookup text,
-                                                          show_code boolean,
-                                                          show_label boolean,
-                                                          variadic links text[]) returns text as $$
+            create or replace function _lookup.lookup_label(code text,
+                                                              lookup text,
+                                                              show_code boolean,
+                                                              show_label boolean,
+                                                              variadic links text[]) returns text as $$
             declare
                 link_name text;
                 link_index int = 0;
@@ -204,21 +175,21 @@ public class LookupExtension implements Extension {
                 result text;
 
             begin
-                from_clause := '"%1$s"."LookupValue" v0 '
-                            || 'join "%1$s"."Lookup" lookup '
+                from_clause := '_lookup."LookupValue" v0 '
+                            || 'join _lookup."Lookup" lookup '
                             || 'on (v0.lookup_id=lookup._id and lookup.name=''' || lookup || ''')';
 
                 foreach link_name in array links loop
                     -- source side
-                    from_clause := from_clause || ' join "%1$s"."LookupValueLink" lk'  || link_index
-                                               || ' on (v' || link_index  || '._id=lk' || link_index
-                                               || '.source_value_id and ' || 'lk'      || link_index
+                    from_clause := from_clause || ' join _lookup."LookupValueLink" lk' || link_index
+                                               || ' on (v' || link_index  || '._id=lk'   || link_index
+                                               || '.source_value_id and ' || 'lk'        || link_index
                                                || '.name=''' || link_name || ''')';
 
                     link_index := link_index + 1;
 
                     -- target side
-                    from_clause := from_clause || ' join "%1$s"."LookupValue" v'      || link_index
+                    from_clause := from_clause || ' join _lookup."LookupValue" v'   || link_index
                                                || ' on (v' || link_index || '._id=lk' || (link_index - 1)
                                                || '.target_value_id)';
                 end loop;
@@ -238,7 +209,7 @@ public class LookupExtension implements Extension {
                 execute query into result;
                 return result;
             end;
-            $$ language plpgsql immutable;""".formatted(schema));
+            $$ language plpgsql immutable;""");
 
         c.commit();
       } catch (SQLException e) {
@@ -248,10 +219,10 @@ public class LookupExtension implements Extension {
       try (Connection c = db.pooledConnection()) {
         // function to find value from lookups
         c.createStatement().executeUpdate("""
-            create or alter function "%1$s".lookup_label0(@Code      nvarchar(max),
-                                                          @Lookup    nvarchar(max),
-                                                          @ShowCode  bit,
-                                                          @ShowLabel bit) returns nvarchar(max) as
+            create or alter function _lookup.lookup_label0(@Code      nvarchar(max),
+                                                             @Lookup    nvarchar(max),
+                                                             @ShowCode  bit,
+                                                             @ShowLabel bit) returns nvarchar(max) as
             begin
               declare @Result nvarchar(max);
               select @Result=(iif(coalesce(@ShowCode, 0)=coalesce(@ShowLabel, 0),
@@ -259,19 +230,18 @@ public class LookupExtension implements Extension {
                                   iif(coalesce(@ShowCode, 0)=1,
                                       code,
                                       label)))
-                from "%1$s"."LookupValue" v
-                join "%1$s"."Lookup"      l on v.lookup_id=l._id
+                from _lookup."LookupValue" v
+                join _lookup."Lookup"      l on v.lookup_id=l._id
               where l.name=@Lookup and v.code=@Code;
               return @Result;
-            end;
-            """.formatted(schema));
+            end;""");
 
         c.createStatement().executeUpdate("""
-            create or alter function "%1$s".lookup_label1(@Code      nvarchar(max),
-                                                          @Lookup    nvarchar(max),
-                                                          @ShowCode  bit,
-                                                          @ShowLabel bit,
-                                                          @Link1     nvarchar(max)) returns nvarchar(max) as
+            create or alter function _lookup.lookup_label1(@Code      nvarchar(max),
+                                                             @Lookup    nvarchar(max),
+                                                             @ShowCode  bit,
+                                                             @ShowLabel bit,
+                                                             @Link1     nvarchar(max)) returns nvarchar(max) as
             begin
               declare @LinkCursor Cursor;
               declare @Result nvarchar(max);
@@ -282,24 +252,23 @@ public class LookupExtension implements Extension {
                                       v1.code,
                                       v1.label)))
 
-                from "%1$s"."LookupValue" v0
-                join "%1$s"."Lookup" lookup on (v0.lookup_id=lookup._id and lookup.name=@Lookup)
+                from _lookup."LookupValue" v0
+                join _lookup."Lookup" lookup on (v0.lookup_id=lookup._id and lookup.name=@Lookup)
 
-                join "%1$s"."LookupValueLink" lk0 on (v0._id=lk0.source_value_id and lk0.name=@Link1)
-                join "%1$s"."LookupValue" v1 on v1._id=lk0.target_value_id
+                join _lookup."LookupValueLink" lk0 on (v0._id=lk0.source_value_id and lk0.name=@Link1)
+                join _lookup."LookupValue" v1 on v1._id=lk0.target_value_id
 
                where v0.code=@Code;
               return @Result;
-            end;
-            """.formatted(schema));
+            end;""");
 
         c.createStatement().executeUpdate("""
-            create or alter function "%1$s".lookup_label2(@Code      nvarchar(max),
-                                                          @Lookup    nvarchar(max),
-                                                          @ShowCode  bit,
-                                                          @ShowLabel bit,
-                                                          @Link1     nvarchar(max),
-                                                          @Link2     nvarchar(max)) returns nvarchar(max) as
+            create or alter function _lookup.lookup_label2(@Code      nvarchar(max),
+                                                             @Lookup    nvarchar(max),
+                                                             @ShowCode  bit,
+                                                             @ShowLabel bit,
+                                                             @Link1     nvarchar(max),
+                                                             @Link2     nvarchar(max)) returns nvarchar(max) as
             begin
               declare @LinkCursor Cursor;
               declare @Result nvarchar(max);
@@ -310,22 +279,21 @@ public class LookupExtension implements Extension {
                                       v2.code,
                                       v2.label)))
 
-                from "%1$s"."LookupValue" v0
-                join "%1$s"."Lookup" lookup on (v0.lookup_id=lookup._id and lookup.name=@Lookup)
+                from _lookup."LookupValue" v0
+                join _lookup."Lookup" lookup on (v0.lookup_id=lookup._id and lookup.name=@Lookup)
 
-                join "%1$s"."LookupValueLink" lk0 on (v0._id=lk0.source_value_id and lk0.name=@Link1)
-                join "%1$s"."LookupValue" v1 on v1._id=lk0.target_value_id
+                join _lookup."LookupValueLink" lk0 on (v0._id=lk0.source_value_id and lk0.name=@Link1)
+                join _lookup."LookupValue" v1 on v1._id=lk0.target_value_id
 
-                join "%1$s"."LookupValueLink" lk1 on (v1._id=lk1.source_value_id and lk1.name=@Link2)
-                join "%1$s"."LookupValue" v2 on v2._id=lk1.target_value_id
+                join _lookup."LookupValueLink" lk1 on (v1._id=lk1.source_value_id and lk1.name=@Link2)
+                join _lookup."LookupValue" v2 on v2._id=lk1.target_value_id
 
                where v0.code=@Code;
               return @Result;
-            end;
-            """.formatted(schema));
+            end;""");
 
         c.createStatement().executeUpdate("""
-            create or alter function "%1$s".lookup_label3(@Code      nvarchar(max),
+            create or alter function _lookup.lookup_label3(@Code      nvarchar(max),
                                                           @Lookup    nvarchar(max),
                                                           @ShowCode  bit,
                                                           @ShowLabel bit,
@@ -342,25 +310,24 @@ public class LookupExtension implements Extension {
                                       v3.code,
                                       v3.label)))
 
-                from "%1$s"."LookupValue" v0
-                join "%1$s"."Lookup" lookup on (v0.lookup_id=lookup._id and lookup.name=@Lookup)
+                from _lookup."LookupValue" v0
+                join _lookup."Lookup" lookup on (v0.lookup_id=lookup._id and lookup.name=@Lookup)
 
-                join "%1$s"."LookupValueLink" lk0 on (v0._id=lk0.source_value_id and lk0.name=@Link1)
-                join "%1$s"."LookupValue" v1 on v1._id=lk0.target_value_id
+                join _lookup."LookupValueLink" lk0 on (v0._id=lk0.source_value_id and lk0.name=@Link1)
+                join _lookup."LookupValue" v1 on v1._id=lk0.target_value_id
 
-                join "%1$s"."LookupValueLink" lk1 on (v1._id=lk1.source_value_id and lk1.name=@Link2)
-                join "%1$s"."LookupValue" v2 on v2._id=lk1.target_value_id
+                join _lookup."LookupValueLink" lk1 on (v1._id=lk1.source_value_id and lk1.name=@Link2)
+                join _lookup."LookupValue" v2 on v2._id=lk1.target_value_id
 
-                join "%1$s"."LookupValueLink" lk2 on (v2._id=lk2.source_value_id and lk2.name=@Link3)
-                join "%1$s"."LookupValue" v3 on v3._id=lk2.target_value_id
+                join _lookup."LookupValueLink" lk2 on (v2._id=lk2.source_value_id and lk2.name=@Link3)
+                join _lookup."LookupValue" v3 on v3._id=lk2.target_value_id
 
                where v0.code=@Code;
               return @Result;
-            end;
-            """.formatted(schema));
+            end;""");
 
         c.createStatement().executeUpdate("""
-            create or alter function "%1$s".lookup_label4(@Code      nvarchar(max),
+            create or alter function _lookup.lookup_label4(@Code      nvarchar(max),
                                                           @Lookup    nvarchar(max),
                                                           @ShowCode  bit,
                                                           @ShowLabel bit,
@@ -378,28 +345,27 @@ public class LookupExtension implements Extension {
                                       v4.code,
                                       v4.label)))
 
-                from "%1$s"."LookupValue" v0
-                join "%1$s"."Lookup" lookup on (v0.lookup_id=lookup._id and lookup.name=@Lookup)
+                from _lookup."LookupValue" v0
+                join _lookup."Lookup" lookup on (v0.lookup_id=lookup._id and lookup.name=@Lookup)
 
-                join "%1$s"."LookupValueLink" lk0 on (v0._id=lk0.source_value_id and lk0.name=@Link1)
-                join "%1$s"."LookupValue" v1 on v1._id=lk0.target_value_id
+                join _lookup."LookupValueLink" lk0 on (v0._id=lk0.source_value_id and lk0.name=@Link1)
+                join _lookup."LookupValue" v1 on v1._id=lk0.target_value_id
 
-                join "%1$s"."LookupValueLink" lk1 on (v1._id=lk1.source_value_id and lk1.name=@Link2)
-                join "%1$s"."LookupValue" v2 on v2._id=lk1.target_value_id
+                join _lookup."LookupValueLink" lk1 on (v1._id=lk1.source_value_id and lk1.name=@Link2)
+                join _lookup."LookupValue" v2 on v2._id=lk1.target_value_id
 
-                join "%1$s"."LookupValueLink" lk2 on (v2._id=lk2.source_value_id and lk2.name=@Link3)
-                join "%1$s"."LookupValue" v3 on v3._id=lk2.target_value_id
+                join _lookup."LookupValueLink" lk2 on (v2._id=lk2.source_value_id and lk2.name=@Link3)
+                join _lookup."LookupValue" v3 on v3._id=lk2.target_value_id
 
-                join "%1$s"."LookupValueLink" lk3 on (v3._id=lk3.source_value_id and lk3.name=@Link4)
-                join "%1$s"."LookupValue" v4 on v4._id=lk3.target_value_id
+                join _lookup."LookupValueLink" lk3 on (v3._id=lk3.source_value_id and lk3.name=@Link4)
+                join _lookup."LookupValue" v4 on v4._id=lk3.target_value_id
 
                where v0.code=@Code;
               return @Result;
-            end;
-            """.formatted(schema));
+            end;""");
 
         c.createStatement().executeUpdate("""
-            create or alter function "%1$s".lookup_label5(@Code      nvarchar(max),
+            create or alter function _lookup.lookup_label5(@Code      nvarchar(max),
                                                           @Lookup    nvarchar(max),
                                                           @ShowCode  bit,
                                                           @ShowLabel bit,
@@ -418,28 +384,27 @@ public class LookupExtension implements Extension {
                                       v5.code,
                                       v5.label)))
 
-                from "%1$s"."LookupValue" v0
-                join "%1$s"."Lookup" lookup on (v0.lookup_id=lookup._id and lookup.name=@Lookup)
+                from _lookup."LookupValue" v0
+                join _lookup."Lookup" lookup on (v0.lookup_id=lookup._id and lookup.name=@Lookup)
 
-                join "%1$s"."LookupValueLink" lk0 on (v0._id=lk0.source_value_id and lk0.name=@Link1)
-                join "%1$s"."LookupValue" v1 on v1._id=lk0.target_value_id
+                join _lookup."LookupValueLink" lk0 on (v0._id=lk0.source_value_id and lk0.name=@Link1)
+                join _lookup."LookupValue" v1 on v1._id=lk0.target_value_id
 
-                join "%1$s"."LookupValueLink" lk1 on (v1._id=lk1.source_value_id and lk1.name=@Link2)
-                join "%1$s"."LookupValue" v2 on v2._id=lk1.target_value_id
+                join _lookup."LookupValueLink" lk1 on (v1._id=lk1.source_value_id and lk1.name=@Link2)
+                join _lookup."LookupValue" v2 on v2._id=lk1.target_value_id
 
-                join "%1$s"."LookupValueLink" lk2 on (v2._id=lk2.source_value_id and lk2.name=@Link3)
-                join "%1$s"."LookupValue" v3 on v3._id=lk2.target_value_id
+                join _lookup."LookupValueLink" lk2 on (v2._id=lk2.source_value_id and lk2.name=@Link3)
+                join _lookup."LookupValue" v3 on v3._id=lk2.target_value_id
 
-                join "%1$s"."LookupValueLink" lk3 on (v3._id=lk3.source_value_id and lk3.name=@Link4)
-                join "%1$s"."LookupValue" v4 on v4._id=lk3.target_value_id
+                join _lookup."LookupValueLink" lk3 on (v3._id=lk3.source_value_id and lk3.name=@Link4)
+                join _lookup."LookupValue" v4 on v4._id=lk3.target_value_id
 
-                join "%1$s"."LookupValueLink" lk4 on (v4._id=lk4.source_value_id and lk4.name=@Link5)
-                join "%1$s"."LookupValue" v5 on v5._id=lk4.target_value_id
+                join _lookup."LookupValueLink" lk4 on (v4._id=lk4.source_value_id and lk4.name=@Link5)
+                join _lookup."LookupValue" v5 on v5._id=lk4.target_value_id
 
                where v0.code=@Code;
               return @Result;
-            end;
-            """.formatted(schema));
+            end;""");
 
         c.commit();
       } catch (SQLException e) {
@@ -497,18 +462,12 @@ public class LookupExtension implements Extension {
                                               source_group:source."group",
                                        source_display_name:source.display_name,
                                         source_description:source.description,
-                                       
-                                                 link_name:link.name,
-                                         link_display_name:link.display_name,
-                                       
                                                target_name:target.name
                                        
                                   from source:_lookup.Lookup
                              left join   link:_lookup.LookupLink on link.source_lookup_id=source._id
                              left join target:_lookup.Lookup     on target._id=link.target_lookup_id""")) {
         record Link(String source,
-                    String name,
-                    String displayName,
                     String target) {}
         Map<String, Lookup> lookups = new HashMap<>();
         List<Link> links = new ArrayList<>();
@@ -524,18 +483,15 @@ public class LookupExtension implements Extension {
                                                new HashMap<>(),
                                                new HashMap<>()));
           }
-          String linkName = rs.value("link_name");
-          if (linkName != null) {
-            links.add(new Link(lookupName,
-                               linkName,
-                               rs.value("link_display_name"),
-                               rs.value("target_name")));
+          String targetName = rs.value("target_name");
+          if (targetName != null) {
+            links.add(new Link(lookupName, targetName));
           }
         }
         for (Link link: links) {
           Lookup source = lookups.get(link.source);
           Lookup target = lookups.get(link.target);
-          source.links().add(new LookupLink(link.name, link.displayName, target));
+          source.links().add(target);
         }
 
         /*
@@ -651,7 +607,7 @@ public class LookupExtension implements Extension {
                     where source_lookup_id=u'""" + lookupId + "'");
 
           if (lookup.links() != null) {
-            for (LookupLink link: lookup.links()) {
+            for (Lookup link: lookup.links()) {
               saveLookupLink(lookupId, link);
             }
           }
@@ -693,31 +649,18 @@ public class LookupExtension implements Extension {
                        .add("description", lookup.description())
                        .add("grp",         lookup.group()));
           if (lookup.links() != null) {
-            for (LookupLink link: lookup.links()) {
+            for (Lookup link: lookup.links()) {
               con.exec("""
-                       insert into _lookup.LookupLink(_id,      name, display_name, source_lookup_id, target_lookup_id)
-                                               values(newid(), @name, @displayName, @sourceId,        @targetId)""",
+                       insert into _lookup.LookupLink(_id,     source_lookup_id, target_lookup_id)
+                                               values(newid(), @sourceId,        @targetId)""",
                        new QueryParams()
-                           .add("name",        link.name())
-                           .add("displayName", link.displayName())
                            .add("sourceId",    lookupId)
-                           .add("targetId",    link.target().id()));
+                           .add("targetId",    link.id()));
             }
           }
           if (lookup.values() != null) {
             for (LookupValue value: lookup.values().values()) {
               saveLookupValue(lookupId, value);
-//              con.exec("""
-//                       insert into _lookup.LookupValue(_id,     lookup_id,  code, alt_code1, alt_code2,  label,  description,  lang)
-//                                                values(newid(), @lookupId, @code, @altCode1, @altCode2, @label, @description, @lang)""",
-//                       new QueryParams()
-//                           .add("lookupId",    lookupId)
-//                           .add("code",        value.code())
-//                           .add("altCode1",    value.altCode1())
-//                           .add("altCode2",    value.altCode2())
-//                           .add("label",       value.label())
-//                           .add("description", value.description())
-//                           .add("lang",        value.lang() == null ? "en" : value.lang()));
             }
           }
         }
@@ -727,40 +670,38 @@ public class LookupExtension implements Extension {
     return lookup.id();
   }
 
-  public void saveLookupLink(UUID lookupId, LookupLink link) {
+  public void saveLookupLink(UUID lookupId, Lookup link) {
     try (EsqlConnection con = db.esql();
          Result rs = con.exec("""
                               select ln._id
                                 from ln:_lookup.LookupLink
                                 join lk:_lookup.Lookup on lk._id=ln.source_lookup_id
-                                                      and lk._id=@lookupId
-                               where ln.name=@linkName
+                                                      and lk._id=@sourceId
+                               where ln.target_lookup_id=@targetId
                               """,
                               new QueryParams()
-                                 .add("lookupId", lookupId)
-                                 .add("linkName", link.name()))) {
+                                 .add("sourceId", lookupId)
+                                 .add("targetId", link.id()))) {
       if (rs.toNext()) {
         UUID linkId = rs.value(1);
         con.exec("""
                  update ln
                    from ln:_lookup.LookupLink
                     set _version=2,
-                        display_name=@displayName,
                         target_lookup_id=@targetId
                   where ln._id=@linkId""",
                  new QueryParams()
-                     .add("displayName", link.displayName())
-                     .add("linkId",      linkId)
-                     .add("targetId",    loadLookup(link.target().name()).id()));
+                     .add("linkId",   linkId)
+                     .add("targetId", link.id()));
+//                     .add("targetId", loadLookup(link.name()).id()));
       } else {
         con.exec("""
-                 insert into _lookup.LookupLink(_id,     _version,  name, display_name, source_lookup_id, target_lookup_id)
-                                         values(newid(), 1,        @name, @displayName, @sourceId,        @targetId)""",
+                 insert into _lookup.LookupLink(_id,     _version, source_lookup_id, target_lookup_id)
+                                         values(newid(), 1,        @sourceId,        @targetId)""",
                  new QueryParams()
-                     .add("name",        link.name())
-                     .add("displayName", link.displayName())
-                     .add("sourceId",    lookupId)
-                     .add("targetId",    loadLookup(link.target().name()).id()));
+                     .add("sourceId", lookupId)
+                     .add("targetId", link.id()));
+//                     .add("targetId", loadLookup(link.target().name()).id()));
       }
     }
   }
@@ -783,12 +724,12 @@ public class LookupExtension implements Extension {
         con.exec("""
                  update lv
                    from lv:_lookup.LookupValue
-                    set _version=2,
-                        alt_code1=@altCode1,
-                        alt_code2=@altCode2,
-                        label=@label,
+                    set _version   =2,
+                        alt_code1  =@altCode1,
+                        alt_code2  =@altCode2,
+                        label      =@label,
                         description=@description,
-                        lang=@lang
+                        lang       =@lang
                   where _id=@valueId""",
                  new QueryParams()
                      .add("valueId",     valueId)
@@ -836,8 +777,6 @@ public class LookupExtension implements Extension {
       }
     }
   }
-
-  private String schema;
 
   private Map<String, Lookup> lookupsCache = null;