+++ src/main/java/ma/vi/esql/lookup/LookupLink.java
@@ -1,8 +1,0 @@
-package ma.vi.esql.lookup;
-
-/**
- * @author Vikash Madhow (vikash.madhow@gmail.com)
- */
-public record LookupLink(String name,
-                         String displayName,
-                         Lookup target) {}